Add files via upload
The quality of the data, the selection of the hyperparameters, and other variables will all have an impact on how accurate the models are. By contrasting the accuracy score and the categorization report, you may assess how accurate the two models are.

Naive Bayes is a general-purpose, quick, and easy method that performs well with high-dimensional data. It is based on the Bayes theorem and makes the strong assumption that the features are independent, which is frequently true in real-world data. However, SVM is a more robust and adaptable technique that is effective with both linear and non-linear data. It looks for the hyperplane with the largest margin between the classes, and it can be regularized to prevent overfitting.